+++ tests.ps1
@@ -1,0 +1,8 @@
+Import-Module Pester
+
+$script:moduleRoot = Split-Path -Path $MyInvocation.MyCommand.Path
+
+# Dot source functions
+"$script:moduleRoot\functions\*.ps1" | Resolve-Path | %{. $_.ProviderPath}
+write-host 'help!'
+Invoke-Pester tests